Refactor Effectiveness to eliminate circular dependencies.
