+++ tslint.json
@@ -34,7 +34,7 @@
     "no-shadowed-variable": true,
     "no-trailing-whitespace": true,
     "no-unused-expression": true,
-    "no-unused-variable": [true, "check-parameters", {"ignore-pattern": "^_"}],
+    "no-unused-variable": [true, {"ignore-pattern": "^_"}],
     "no-use-before-declare": true,
     "no-var-keyword": true,
     "one-line": [
